[commitgraph] refactor
